# Improve `scan` in cupy/core/_routines_math.pyx.