--- a/cupy/core/_routines_math.pyx
+++ b/cupy/core/_routines_math.pyx
@@ -147,6 +147,323 @@ cdef ndarray _ndarray_clip(ndarray self, a_min, a_max, out):
 
 # private/internal
 
+_op_char = {scan_op.SCAN_SUM: '+', scan_op.SCAN_PROD: '*'}
+_identity = {scan_op.SCAN_SUM: 0, scan_op.SCAN_PROD: 1}
+
+
+@cupy._util.memoize(for_each_device=True)
+def _cupy_bsum_shfl(op, block_size, warp_size=32):
+    """Returns a kernel that computes the sum/prod of each block.
+
+    Args:
+        op (int): Operation type. SCAN_SUM or SCAN_PROD.
+        block_size (int): Block size.
+        warp_size (int); Warp size.
+
+    Returns:
+        cupy.ElementwiseKernel
+
+    Example:
+        a = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+        _cupy_bsum(op=SCAN_SUM, block_size=4)(a, a.size, b, ...)
+        b == [10, 26, 19]
+
+    Note:
+        This uses warp shuffle functions to exchange data in a warp.
+        See the link below for details about warp shuffle functions.
+        https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#warp-shuffle-functions
+    """
+    in_params = 'raw T a'
+    out_params = 'raw O b'
+    loop_prep = string.Template("""
+        __shared__ O smem[${block_size} / ${warp_size}];
+        const int n_warp = ${block_size} / ${warp_size};
+        const int warp_id = threadIdx.x / ${warp_size};
+        const int lane_id = threadIdx.x % ${warp_size};
+    """).substitute(block_size=block_size, warp_size=warp_size)
+    loop_body = string.Template("""
+        O x = ${identity};
+        if (i < a.size()) x = a[i];
+        for (int j = 1; j < ${warp_size}; j *= 2) {
+            x ${op}= __shfl_xor_sync(0xffffffff, x, j);
+        }
+        if (lane_id == 0) smem[warp_id] = x;
+        __syncthreads();
+        if (warp_id == 0) {
+            x = ${identity};
+            if (lane_id < n_warp) x = smem[lane_id];
+            for (int j = 1; j < n_warp; j *= 2) {
+                x ${op}= __shfl_xor_sync(0xffffffff, x, j);
+            }
+            int block_id = i / ${block_size};
+            if (lane_id == 0) b[block_id] = x;
+        }
+    """).substitute(block_size=block_size, warp_size=warp_size,
+                    op=_op_char[op], identity=_identity[op])
+    return cupy.ElementwiseKernel(in_params, out_params, loop_body,
+                                  'cupy_bsum_shfl', loop_prep=loop_prep)
+
+
+@cupy._util.memoize(for_each_device=True)
+def _cupy_bsum_smem(op, block_size, warp_size=32):
+    """Returns a kernel that computes the sum/prod of each block.
+
+    Args:
+        op (int): Operation type. SCAN_SUM or SCAN_PROD.
+        block_size (int): Block size.
+        warp_size (int); Warp size.
+
+    Returns:
+        cupy.ElementwiseKernel
+
+    Example:
+        a = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+        _cupy_bsum(op=SCAN_SUM, block_size=4)(a, a.size, b, ...)
+        b == [10, 26, 19]
+
+    Note:
+        This uses shared memory to exchange data in a warp.
+    """
+    in_params = 'raw T a'
+    out_params = 'raw O b'
+    loop_prep = string.Template("""
+        __shared__ O smem1[${block_size}];
+        __shared__ O smem2[${warp_size}];
+        const int n_warp = ${block_size} / ${warp_size};
+        const int warp_id = threadIdx.x / ${warp_size};
+        const int lane_id = threadIdx.x % ${warp_size};
+    """).substitute(block_size=block_size, warp_size=warp_size)
+    loop_body = string.Template("""
+        O x = ${identity};
+        if (i < a.size()) x = a[i];
+        for (int j = 1; j < ${warp_size}; j *= 2) {
+            smem1[threadIdx.x] = x;          __syncwarp();
+            x ${op}= smem1[threadIdx.x ^ j]; __syncwarp();
+        }
+        if (lane_id == 0) smem2[warp_id] = x;
+        __syncthreads();
+        if (warp_id == 0) {
+            x = ${identity};
+            if (lane_id < n_warp) x = smem2[lane_id];
+            for (int j = 1; j < n_warp; j *= 2) {
+                smem2[lane_id] = x;          __syncwarp();
+                x ${op}= smem2[lane_id ^ j]; __syncwarp();
+            }
+            int block_id = i / ${block_size};
+            if (lane_id == 0) b[block_id] = x;
+        }
+    """).substitute(block_size=block_size, warp_size=warp_size,
+                    op=_op_char[op], identity=_identity[op])
+    return cupy.ElementwiseKernel(in_params, out_params, loop_body,
+                                  'cupy_bsum_smem', loop_prep=loop_prep)
+
+
+@cupy._util.memoize(for_each_device=True)
+def _cupy_scan_naive(op, block_size, warp_size=32):
+    """Returns a kernel to compute an inclusive scan.
+
+    It first performs an inclusive scan in each block and the add the
+    scan results for the sum/prod of each block.
+
+    Args:
+        op (int): Operation type. SCAN_SUM or SCAN_PROD.
+        block_size (int): Block size.
+        warp_size (int); Warp size.
+
+    Returns:
+        cupy.ElementwiseKernel
+
+    Example:
+        b = [10, 36, 55]
+        a = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+        _cupy_scan(op=SCAN_SUM, block_size=4)(b, a.size, a, out, ...)
+        out == [1, 3, 6, 10, 15, 21, 28, 36, 45, 55]
+
+    Note:
+        This uses a kind of method called "Naive Parallel Scan" for inclusive
+        scan in each block. See below for details about "Naive Parallel Scan".
+        https://developer.nvidia.com/gpugems/gpugems3/part-vi-gpu-computing/chapter-39-parallel-prefix-sum-scan-cuda
+    """
+    in_params = 'raw O b'
+    out_params = 'raw T a, raw O out'
+    loop_prep = string.Template("""
+        __shared__ O smem1[${block_size}];
+        __shared__ O smem2[${warp_size}];
+        const int n_warp = ${block_size} / ${warp_size};
+        const int warp_id = threadIdx.x / ${warp_size};
+        const int lane_id = threadIdx.x % ${warp_size};
+    """).substitute(block_size=block_size, warp_size=warp_size)
+    loop_body = string.Template("""
+        O x = ${identity};
+        if (i < a.size()) x = a[i];
+        for (int j = 1; j < ${warp_size}; j *= 2) {
+            smem1[threadIdx.x] = x;  __syncwarp();
+            if (lane_id - j >= 0) x ${op}= smem1[threadIdx.x - j];
+            __syncwarp();
+        }
+        if (lane_id == ${warp_size} - 1) smem2[warp_id] = x;
+        __syncthreads();
+        if (warp_id == 0) {
+            O y = ${identity};
+            if (lane_id < n_warp) y = smem2[lane_id];
+            for (int j = 1; j < n_warp; j *= 2) {
+                smem2[lane_id] = y;  __syncwarp();
+                if (lane_id - j >= 0) y ${op}= smem2[lane_id - j];
+                __syncwarp();
+            }
+            smem2[lane_id] = y;
+        }
+        __syncthreads();
+        if (warp_id > 0) x ${op}= smem2[warp_id - 1];
+        int block_id = i / ${block_size};
+        if (block_id > 0) x ${op}= b[block_id - 1];
+        if (i < a.size()) out[i] = x;
+    """).substitute(block_size=block_size, warp_size=warp_size,
+                    op=_op_char[op], identity=_identity[op])
+    return cupy.ElementwiseKernel(in_params, out_params, loop_body,
+                                  'cupy_scan_naive', loop_prep=loop_prep)
+
+
+@cupy._util.memoize(for_each_device=True)
+def _cupy_scan_btree(op, block_size, warp_size=32):
+    """Returns a kernel to compute an inclusive scan.
+
+    It first performs an inclusive scan in each block and the add the
+    scan results for the sum/prod of each block.
+
+    Args:
+        op (int): Operation type. SCAN_SUM or SCAN_PROD.
+        block_size (int): Block size.
+        warp_size (int); Warp size.
+
+    Returns:
+        cupy.ElementwiseKernel
+
+    Example:
+        b = [10, 36, 55]
+        a = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+        _cupy_scan(op=SCAN_SUM, block_size=4)(b, a.size, a, out, ...)
+        out == [1, 3, 6, 10, 15, 21, 28, 36, 45, 55]
+
+    Note:
+        This usea a kind of method called "Work-Efficenet Parallel Scan" for
+        inclusive scan in each block. See below link for details about
+        "Work-Efficent Parallel Scan".
+        https://developer.nvidia.com/gpugems/gpugems3/part-vi-gpu-computing/chapter-39-parallel-prefix-sum-scan-cuda
+    """
+    in_params = 'raw O b'
+    out_params = 'raw T a, raw O out'
+    loop_prep = string.Template("""
+        __shared__ O smem0[${block_size} + 1];
+        O *smem1 = smem0 + 1;
+        __shared__ O smem2[${warp_size}];
+        const int n_warp = ${block_size} / ${warp_size};
+        const int warp_id = threadIdx.x / ${warp_size};
+        const int lane_id = threadIdx.x % ${warp_size};
+        if (threadIdx.x == 0) smem0[0] = ${identity};
+    """).substitute(block_size=block_size, warp_size=warp_size,
+                    identity=_identity[op])
+    loop_body = string.Template("""
+        O x = ${identity};
+        if (i < a.size()) x = a[i];
+        for (int j = 1; j < ${warp_size}; j *= 2) {
+            smem1[threadIdx.x] = x;  __syncwarp();
+            if (lane_id % (2*j) == (2*j)-1) {
+                x ${op}= smem1[threadIdx.x - j];
+            }
+            __syncwarp();
+        }
+        smem1[threadIdx.x] = x;
+        __syncthreads();
+        if (warp_id == 0) {
+            O y = ${identity};
+            if (lane_id < n_warp) {
+                y = smem0[${warp_size} * (lane_id + 1)];
+            }
+            for (int j = 1; j < n_warp; j *= 2) {
+                smem2[lane_id] = y;  __syncwarp();
+                if (lane_id % (2*j) == (2*j)-1) {
+                    y ${op}= smem2[lane_id - j];
+                }
+                __syncwarp();
+            }
+            for (int j = n_warp / 4; j > 0; j /= 2) {
+                smem2[lane_id] = y; __syncwarp();
+                if ((lane_id % (2*j) == j-1) && (lane_id >= 2*j)) {
+                    y ${op}= smem2[lane_id - j];
+                }
+                __syncwarp();
+            }
+            if (lane_id < n_warp) {
+                smem0[${warp_size} * (lane_id + 1)] = y;
+            }
+        }
+        __syncthreads();
+        x = smem0[threadIdx.x];
+        for (int j = ${warp_size} / 2; j > 0; j /= 2) {
+            if (lane_id % (2*j) == j) {
+                x ${op}= smem0[threadIdx.x - j];
+            }
+            __syncwarp();
+            smem0[threadIdx.x] = x;  __syncwarp();
+        }
+        __syncthreads();
+        x = smem1[threadIdx.x];
+        int block_id = i / ${block_size};
+        if (block_id > 0) x ${op}= b[block_id - 1];
+        if (i < a.size()) out[i] = x;
+    """).substitute(block_size=block_size, warp_size=warp_size,
+                    op=_op_char[op], identity=_identity[op])
+    return cupy.ElementwiseKernel(in_params, out_params, loop_body,
+                                  'cupy_scan_btree', loop_prep=loop_prep)
+
+
+cdef ndarray scan(ndarray a, op, dtype=None, ndarray out=None):
+    """Return the prefix sum(scan) of the elements.
+
+    Args:
+        a (cupy.ndarray): input array.
+        out (cupy.ndarray): Alternative output array in which to place
+         the result. The same size and same type as the input array(a).
+
+    Returns:
+        cupy.ndarray: A new array holding the result is returned.
+
+    """
+    if a._shape.size() != 1:
+        raise TypeError('Input array should be 1D array.')
+
+    if out is None:
+        if dtype is None:
+            dtype = a.dtype
+        out = _ndarray_init(a._shape, dtype)
+    else:
+        if a.size != out.size:
+            raise ValueError('Provided out is the wrong size')
+
+    block_size = 512
+    warp_size = 32
+    if out.dtype.char in 'iIlLqQfd':
+        bsum_kernel = _cupy_bsum_shfl(op, block_size, warp_size)
+    else:
+        bsum_kernel = _cupy_bsum_smem(op, block_size, warp_size)
+    if out.dtype.char in 'fdFD':
+        scan_kernel = _cupy_scan_btree(op, block_size, warp_size)
+    else:
+        scan_kernel = _cupy_scan_naive(op, block_size, warp_size)
+    b_size = (a.size + block_size - 1) // block_size
+    b = cupy.empty((b_size,), dtype=out.dtype)
+    size = b.size * block_size
+
+    if a.size > block_size:
+        bsum_kernel(a, b, size=size, block_size=block_size)
+        scan(b, op, dtype=out.dtype, out=b)
+        scan_kernel(b, a, out, size=size, block_size=block_size)
+    else:
+        scan_kernel(b, a, out, size=size, block_size=block_size)
+
+    return out
+
 
 @_util.memoize(for_each_device=True)
 def _inclusive_batch_scan_kernel(
@@ -263,85 +580,6 @@ def _inclusive_batch_scan_kernel(
 
 
 @_util.memoize(for_each_device=True)
-def _inclusive_scan_kernel(src_dtype, dtype, block_size, op, src_c_cont,
-                           out_c_cont):
-    """return Prefix Sum(Scan) cuda kernel
-
-    e.g
-    if blocksize * 2 >= len(src)
-    src [1, 2, 3, 4]
-    dst [1, 3, 6, 10]
-
-    if blocksize * 2 < len(src)
-    block_size: 2
-    src [1, 2, 3, 4, 5, 6]
-    dst [1, 3, 6, 10, 5, 11]
-
-    Args:
-        dtype: src, dst array type
-        block_size: block_size
-
-    Returns:
-         cupy.cuda.Function: cuda function
-    """
-
-    name = 'inclusive_scan_kernel'
-    src_dtype = get_typename(src_dtype)
-    dtype = get_typename(dtype)
-    op_char = {scan_op.SCAN_SUM: '+', scan_op.SCAN_PROD: '*'}
-    identity = {scan_op.SCAN_SUM: 0, scan_op.SCAN_PROD: 1}
-    source = string.Template("""
-    extern "C" __global__ void ${name}(
-        const CArray<${src_dtype}, 1, ${src_c_cont}> src,
-        CArray<${dtype}, 1, ${out_c_cont}> dst){
-        long long n = src.size();
-        __shared__ ${dtype} temp[${block_size} * 2];
-        unsigned int thid = threadIdx.x;
-        unsigned int block = 2 * blockIdx.x * blockDim.x;
-
-        unsigned int idx0 = thid + block;
-        unsigned int idx1 = thid + blockDim.x + block;
-
-        temp[thid] = (idx0 < n) ? (${dtype})src[idx0] : (${dtype})${identity};
-        if (idx1 < n) {
-            temp[thid + blockDim.x] = (${dtype}) src[idx1];
-        } else {
-            temp[thid + blockDim.x] = (${dtype}) ${identity};
-        }
-        __syncthreads();
-
-        for(int i = 1; i <= ${block_size}; i <<= 1){
-            int index = (threadIdx.x + 1) * i * 2 - 1;
-            if (index < (${block_size} << 1)){
-                temp[index] ${op}= temp[index - i];
-            }
-            __syncthreads();
-        }
-
-        for(int i = ${block_size} >> 1; i > 0; i >>= 1){
-            int index = (threadIdx.x + 1) * i * 2 - 1;
-            if(index + i < (${block_size} << 1)){
-                temp[index + i] ${op}= temp[index];
-            }
-            __syncthreads();
-        }
-
-        if(idx0 < n){
-            dst[idx0] = temp[thid];
-        }
-        if(idx1 < n){
-            dst[idx1] = temp[thid + blockDim.x];
-        }
-    }
-    """).substitute(name=name, dtype=dtype, block_size=block_size,
-                    src_dtype=src_dtype,
-                    op=op_char[op], identity=identity[op],
-                    src_c_cont=src_c_cont, out_c_cont=out_c_cont)
-    module = compile_with_cache(source)
-    return module.get_function(name)
-
-
-@_util.memoize(for_each_device=True)
 def _add_scan_batch_blocked_sum_kernel(dtype, op, block_size, c_cont):
     name = 'add_scan_blocked_sum_kernel'
     dtype = get_typename(dtype)
@@ -375,70 +613,6 @@ def _add_scan_batch_blocked_sum_kernel(dtype, op, block_size, c_cont):
                     c_cont=c_cont)
     module = compile_with_cache(source)
     return module.get_function(name)
-
-
-@_util.memoize(for_each_device=True)
-def _add_scan_blocked_sum_kernel(dtype, op, c_cont):
-    name = 'add_scan_blocked_sum_kernel'
-    dtype = get_typename(dtype)
-    ops = {scan_op.SCAN_SUM: '+', scan_op.SCAN_PROD: '*'}
-    source = string.Template("""
-    extern "C" __global__ void ${name}(CArray<${dtype}, 1, ${c_cont}> src_dst){
-        long long n = src_dst.size();
-        unsigned int idxBase = (blockDim.x + 1) * (blockIdx.x + 1);
-        unsigned int idxAdded = idxBase + threadIdx.x;
-        unsigned int idxAdd = idxBase - 1;
-
-        if(idxAdded < n){
-            src_dst[idxAdded] ${op}= src_dst[idxAdd];
-        }
-    }
-    """).substitute(name=name, dtype=dtype, op=ops[op], c_cont=c_cont)
-    module = compile_with_cache(source)
-    return module.get_function(name)
-
-
-cdef ndarray scan(ndarray a, op, dtype=None, ndarray out=None):
-    """Return the prefix sum(scan) of the elements.
-
-    Args:
-        a (cupy.ndarray): input array.
-        out (cupy.ndarray): Alternative output array in which to place
-         the result. The same size and same type as the input array(a).
-
-    Returns:
-        cupy.ndarray: A new array holding the result is returned.
-
-    """
-    if a._shape.size() != 1:
-        raise TypeError('Input array should be 1D array.')
-
-    cdef Py_ssize_t block_size = 256
-    if dtype is None:
-        dtype = a.dtype
-    if out is None:
-        out = _ndarray_init(a._shape, dtype)
-    else:
-        if a.size != out.size:
-            raise ValueError('Provided out is the wrong size')
-
-    cdef int src_cont = int(a._c_contiguous)
-    cdef int out_cont = int(out._c_contiguous)
-    kern_scan = _inclusive_scan_kernel(a.dtype, dtype, block_size, op,
-                                       src_cont, out_cont)
-    kern_scan(grid=((a.size - 1) // (2 * block_size) + 1,),
-              block=(block_size,),
-              args=(a, out))
-
-    if (a.size - 1) // (block_size * 2) > 0:
-        blocked_sum = out[block_size * 2 - 1:None:block_size * 2]
-        scan(blocked_sum, op, dtype, blocked_sum)
-        kern_add = _add_scan_blocked_sum_kernel(
-            dtype, op, out_cont)
-        kern_add(grid=((a.size - 1) // (2 * block_size),),
-                 block=(2 * block_size - 1,),
-                 args=(out,))
-    return out
 
 
 cdef ndarray _batch_scan_op(ndarray a, scan_op op, dtype, ndarray out):