This PR improves a function `scan` in cupy/core/_routines_math.pyx.

The current `scan` implementation has a recursive structure, with the 1st kernel performing an inclusive scan in each block and the 2nd kernel adding the base values to the results of 1st kernel. This is not bad, but it has a performance problem in that both kernels cause device memory writes for all array elements.

The `scan` implementation of this PR also has a recursive structure, but the 1st kernel just computes the sum (or prod) of each block, while the 2nd kernel performs the inclusive scan in each block and adds the base values to the results. The advantage of this change is that 1st kernel only needs to write the results of sum (or prod) of each block, thus reducing the amount of writes to device memory.

Below is a performance comparison of `cumsum`.

size | dtype | NumPy | CuPy (master) | CuPy (this PR, https://github.com/cupy/cupy/pull/4366/commits/74aedd2f22ab5a90961e1f2dc8e677a4999d01d9)
--|--|--|--|--
1,000,000 | int32 | 2.982 ms | 0.141 ms | 0.128 ms
1,000,000 | int64 | 2.107 ms | 0.140 ms | 0.114 ms
1,000,000 | float32 | 2.528 ms | 0.092 ms | 0.118 ms
1,000,000 | float64 | 2.569 ms | 0.139 ms | 0.122 ms
10,000,000 | int32 | 40.161 ms | 0.862 ms | 0.652 ms
10,000,000 | int64 | 26.850 ms | 0.937 ms | 0.718 ms
10,000,000 | float32 | 28.207 ms | 0.565 ms | 0.476 ms
10,000,000 | float64 | 31.628 ms | 0.930 ms | 0.728 ms
100,000,000 | int32 | 395.550 ms | 8.268 ms | 6.050 ms
100,000,000 | int64 | 269.081 ms | 8.928 ms | 6.708 ms
100,000,000 | float32 | 280.797 ms | 4.996 ms | 4.282 ms
100,000,000 | float64 | 317.051 ms | 8.588 ms | 6.779 ms

<details>
<summary> Click to see the script </summary>

```python
import numpy
import cupy
import cupyx

n_repeat = 10
n_warmup = 1


def get_time(perf):
    cpu_time = perf.cpu_times.mean()
    gpu_time = perf.gpu_times.mean()
    return max(cpu_time, gpu_time) * 1000  # ms


def test(n, dtype):
    dtype = numpy.dtype(dtype)
    if dtype.char in 'fdFD':
        a = numpy.random.random((n,)).astype(dtype.char.lower()) / (n / 2)
        if dtype.char in 'FD':
            ai = numpy.random.random((n,)).astype(dtype.char.lower()) / (n / 2)
            a = a + 1j * ai
    else:
        a = numpy.random.randint(0, 2, size=(n,)).astype(dtype)
    cp_a = cupy.array(a)
    
    ref = numpy.cumsum(a)
    ret = cupy.cumsum(cp_a)
    if dtype.char not in 'fF':
        cupy.testing.assert_allclose(ret, ref, rtol=1e-5, atol=1e-5)

    ret = []
    perf = cupyx.time.repeat(numpy.cumsum, (a,),
                             n_warmup=n_warmup, n_repeat=n_repeat)
    ret.append(get_time(perf))

    perf = cupyx.time.repeat(cupy.cumsum, (cp_a,),
                             n_warmup=n_warmup, n_repeat=n_repeat)
    ret.append(get_time(perf))
    return ret


dtype_set = ('int32', 'int64', 'float32', 'float64')
n_set = (1000000, 10000000, 100000000)

# cupy.core.set_routine_accelerators(['cub'])

print('size | dtype | NumPy | CuPy')
print('--|--|--|--')
for n in n_set:
    for dtype in dtype_set:
        ret = test(n, dtype)
        print('{} | {} | {:.3f} ms | {:.3f} ms'.
              format(n, dtype, ret[0], ret[1]))
```
</details>