@emcastillo 2020-12-07
Hi, we were discussing about this and @leofang kindly tested this in a HIP platform. 
Unfortunately it didn't work 😣.

This causes the following issue to appear. We need to have this kernel for AMD as well, so we will need to keep the original and add this one only for cuda devices. But for CUDA devices, CUB seems to be faster than this PR. Is there any case where CUB can't be used or this PR shows a significant advantage?
We think that now that CUB is shipped with cupy, users that need performance on these routines can switch to CUB acceleration while we keep a single generic one that is not super-performant, but good enough in the majority of the use cases.

Thanks a lot for the effort you put into this! it really looks amazing