@anaruse 2020-12-07
Thanks for your comment, @emcastillo !

> But for CUDA devices, CUB seems to be faster than this PR. Is there any case where CUB can't be used or this PR shows a significant advantage?

It is true that pure inclusive scan performance is faster with CUB. The CUB inclusive scan is a kind of single kernel implementation. In terms of the amount of device memory access, CUB is theoretically 1.5x faster than this PR.

So why do we need this PR? That is because the scan implementation of this PR can be used to speed-up `nonzero` (and `argwhere` ), that is one of the operations that users often use without knowing it.

The current `nonzero` implementation first performs an inclusive scan and then computes an index array based of the results of scan. With this implementation, you need to write the results of scan to the device memory, regardless of whether you use CUB or not. However, if you can perform inclusive scan and index array computation in the same kernel, you do not need to write the results of inclusive scan to the device memory. This is not possible when using CUB, but can be achieved by taking advantage of the scan implementation of this PR. More specifically, by fusing the 2nd kernel of this PR with the index array computation, you can get the index array without writing the result of the scan back to the device memory. This is a advantage of this PR, see the PR #4367 for more details.