@emcastillo 2020-12-09
Thanks a lot for the detailed explanation.
I see now the potential benefit, so basically you mean that with this PR `nonzero` and `argwhere` will be faster than using CUB because we don't need to fully complete the scan.
If such is the case then this PR is fine :). But I would like to see the result of those two routines with CUB enabled and this PR if you don't mind 🙇‍♂️

We also need to restore the old kernel and add the fallback for hip.