@anaruse 2020-12-09
> Hi, we were discussing about this and @leofang kindly tested this in a HIP platform.
> Unfortunately it didn't work

Could you tell me which part of this PR were problematic in the HIP environment?

This PR uses warp shuffle functions in `_cupy_bsum_shfl`, was that part the problem? If that is the problem, wouldn't using `_cupy_bsum_smem`, which uses shared memory instead of warp shuffle, solve the problem?