@emcastillo 2020-12-09
This one is with this PR included?

If so, this is great!!
probably we want to disable CUB scan when called from nonzero or argwhere :)