@emcastillo 2020-12-09
I think ` __syncwarp();` is not supported by HIP cc @leofang