@anaruse 2020-12-09
> I think __syncwarp(); is not supported by HIP cc @leofang

Isn't there any primitive equivalent to __syncwarp() in HIP?
It's kind of amazing...