@anaruse 2020-12-09
Or maybe it means you don't need __syncwarp() in HIP?
If so, in the case of HIP, maybe you can simply disable the __syncwarp() using a macro or something like that.