@emcastillo 2020-12-09
https://www.olcf.ornl.gov/wp-content/uploads/2019/10/CAAR_HIP_on_Frontier.pdf

` __syncwarp() ` is not supported as per the above slides, so probably this kernel won't work :), or we will need to resort to `__syncthreads()` making it possible slower than the previous implementation?