@anaruse 2020-12-09
One of the reasons we added __syncwarp() in CUDA 9 is the independent thread scheduling that was added from the Volta generation GPUs. Maybe the red GPU does not have this feature? If so, that would explain it.